magika-python-package/pyproject.toml
[tool.poetry]
name = "magika"
version = "0.4.2-dev"
description = "A tool to determine the content type of a file with deep-learning"
authors = ["Yanick Fratantonio <yanickf@google.com>"]
readme = "README.md"
packages = [{include = "magika"}]

[tool.poetry.dependencies]
python = "^3.8,<3.12"
click = "^8.1.3"
tqdm = "^4.64.1"
onnxruntime = "^1.15.1"
numpy = "^1.24.4"
tabulate = "^0.9.0"
python-dotenv = "^1.0.1"
rich = "^13.7.0"

[tool.poetry.scripts]
magika = "magika.cli.magika:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ipython = "^8.10.0"
ruff = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
